+++ feature/login/src/main/kotlin/com/ninecraft/booket/feature/login/client/GoogleLoginClient.kt
@@ -1,4 +1,4 @@
-package com.ninecraft.booket.feature.login
+package com.ninecraft.booket.feature.login.client
 
 import android.content.Context
 import androidx.credentials.CredentialManager
@@ -10,7 +10,7 @@ import androidx.credentials.exceptions.GetCredentialException
 import androidx.credentials.exceptions.NoCredentialException
 import com.google.android.libraries.identity.googleid.GetGoogleIdOption
 import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential
-import com.ninecraft.booket.core.designsystem.R as designR
+import com.ninecraft.booket.core.designsystem.R
 import com.orhanobut.logger.Logger
 import dev.zacsweers.metro.Inject
 
@@ -22,7 +22,7 @@ internal class GoogleLoginClient {
         onSuccess: (String) -> Unit,
         onFailure: (String) -> Unit,
     ) {
-        val credentialManager = CredentialManager.create(context)
+        val credentialManager = CredentialManager.Companion.create(context)
 
         val googleIdOption = GetGoogleIdOption.Builder()
             .setFilterByAuthorizedAccounts(false)
@@ -41,17 +41,17 @@ internal class GoogleLoginClient {
             )
             handleSignIn(result, onSuccess, onFailure, context)
         } catch (e: GetCredentialCancellationException) {
-            Logger.e("Google 로그인 취소됨")
-            onFailure(context.getString(designR.string.unknown_error_message))
+            Logger.e("Google 로그인 취소됨, ${e.message}")
+            onFailure(context.getString(R.string.unknown_error_message))
         } catch (e: NoCredentialException) {
-            Logger.e("Google 계정을 찾을 수 없음")
-            onFailure(context.getString(designR.string.unknown_error_message))
+            Logger.e("Google 계정을 찾을 수 없음, ${e.message}")
+            onFailure(context.getString(R.string.unknown_error_message))
         } catch (e: GetCredentialException) {
-            Logger.e("Google 로그인 실패: ${e.message}")
-            onFailure(context.getString(designR.string.network_error_message))
+            Logger.e("Google 로그인 실패, ${e.message}")
+            onFailure(context.getString(R.string.unknown_error_message))
         } catch (e: Exception) {
             Logger.e("알 수 없는 오류: ${e.message}")
-            onFailure(context.getString(designR.string.unknown_error_message))
+            onFailure(context.getString(R.string.unknown_error_message))
         }
     }
 
@@ -64,19 +64,19 @@ internal class GoogleLoginClient {
         val credential = result.credential
 
         if (credential is CustomCredential &&
-            credential.type == GoogleIdTokenCredential.TYPE_GOOGLE_ID_TOKEN_CREDENTIAL) {
+            credential.type == GoogleIdTokenCredential.Companion.TYPE_GOOGLE_ID_TOKEN_CREDENTIAL) {
             try {
-                val googleIdTokenCredential = GoogleIdTokenCredential.createFrom(credential.data)
+                val googleIdTokenCredential = GoogleIdTokenCredential.Companion.createFrom(credential.data)
                 val idToken = googleIdTokenCredential.idToken
                 Logger.d("Google 로그인 성공: ${googleIdTokenCredential.id}")
                 onSuccess(idToken)
             } catch (e: Exception) {
                 Logger.e("Google ID Token 파싱 실패: ${e.message}")
-                onFailure(context.getString(designR.string.unknown_error_message))
+                onFailure(context.getString(R.string.unknown_error_message))
             }
         } else {
             Logger.e("예상치 못한 credential type: ${credential.type}")
-            onFailure(context.getString(designR.string.unknown_error_message))
+            onFailure(context.getString(R.string.unknown_error_message))
         }
     }
 }
